nsfw-image-detect/requirements.txt
transformers
torch
pillow
gradio
hf_xet #It's not effective ; it even been removed.
fastapi
uvicorn[standard]